+++ packages/orq-rc/RELEASES.md
@@ -7189,3 +7189,13 @@ Based on:
 - [typescript v4.4.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.36 - packages/orq-rc
+
+## 2026-02-23 19:45:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.37 - packages/orq-rc